// frontend/doapet/src/app/models/ong.ts
import { Injectable } from '@angular/core';

@Injectable()
export class Ong {
  name: string = '';
  email: string = '';
  password: string = '';
  phone_number: string = '';
  street: string = '';
  number: string = '';
  neighborhood: string = '';
  city: string = '';
  state: string = '';
  cnpj: string = '';
  pix: string = '';
  president_name: string = '';
  biography: string = '';
  profile_picture: string = '';
}