// frontend/doapet/src/app/models/donor.ts
import { Injectable } from '@angular/core';

@Injectable()
export class Donor {
  name: string = '';
  email: string = '';
  password: string = '';
  phone_number: string = '';
  street: string = '';
  number: string = '';
  neighborhood: string = '';
  city: string = '';
  state: string = '';
  cpf: string = '';
  birthdate: string = '';
  profile_picture: string = '';
}